updated cohortList to sort by date, and moved dispatch to generator
